+++ lib/http1.c
@@ -749,7 +749,7 @@ void reqread_on_read(h2o_socket_t *sock, const char *err)
         conn->_req_entity_reader->handle_incoming_entity(conn);
 }
 
-static int close_idle_connection(h2o_conn_t *_conn)
+static size_t close_idle_connection(h2o_conn_t *_conn)
 {
     struct st_h2o_http1_conn_t *conn = (void *)_conn;
     if (conn->sock->input->size == 0) {
